fixing androidmanifest, Login, Register. +intent
