--- app/src/main/java/com/exless/LoginActivity.kt
@@ -1,13 +1,0 @@
-package com.exless
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-
-class LoginActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_login)
-
-        supportActionBar?.hide()
-    }
-}
