--- app/src/main/java/com/exless/Register_Activity.kt
@@ -1,13 +1,0 @@
-package com.exless
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-
-class Register_Activity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_register)
-
-        supportActionBar?.hide()
-    }
-}
