+++ app/src/main/java/com/exless/view/LoginActivity.kt
@@ -1,0 +1,24 @@
+package com.exless.view
+
+import android.annotation.SuppressLint
+import android.content.Intent
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import android.widget.TextView
+import com.exless.R
+
+class LoginActivity : AppCompatActivity() {
+    @SuppressLint("WrongViewCast")
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_login)
+
+        supportActionBar?.hide()
+
+        val buttonregister = findViewById<TextView>(R.id.tv_toregister)
+        buttonregister.setOnClickListener {
+            val intent = Intent(this, Register_Activity::class.java)
+            startActivity(intent)
+        }
+    }
+}
