+++ app/src/main/java/com/exless/view/Register_Activity.kt
@@ -3,18 +3,18 @@ package com.exless.view
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.widget.Button
+import android.widget.TextView
 import com.exless.R
 
-class onbonding3 : AppCompatActivity() {
+class Register_Activity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_onbonding3)
+        setContentView(R.layout.activity_register)
 
         supportActionBar?.hide()
 
-        val buttomnext = findViewById<Button>(R.id.next)
-        buttomnext.setOnClickListener {
+        val buttonlogin = findViewById<TextView>(R.id.tv_tologin)
+        buttonlogin.setOnClickListener {
             val intent = Intent(this, LoginActivity::class.java)
             startActivity(intent)
         }
