+++ app/src/main/res/drawable/cursorcolor.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android" >
+    <size android:width="3dp" />
+    <solid android:color="#000000"  />
+</shape>
