+++ app/src/main/res/layout/activity_register.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#FFFFFF"
-    tools:context=".LoginActivity">
+    tools:context=".view.LoginActivity">
 
     <TextView
         android:id="@+id/textView"
@@ -71,7 +71,7 @@
         app:layout_constraintTop_toBottomOf="@+id/editTextTextPassword2" />
 
     <TextView
-        android:id="@+id/textView7"
+        android:id="@+id/tv_tologin"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginStart="4dp"
@@ -91,8 +91,8 @@
         android:ems="10"
         android:inputType="text"
         android:padding="10dp"
-        android:text="tes"
         android:textColor="#626262"
+        android:textCursorDrawable="@drawable/cursorcolor"
         app:layout_constraintStart_toStartOf="@+id/textView2"
         app:layout_constraintTop_toBottomOf="@+id/textView2" />
 
@@ -105,40 +105,61 @@
         android:ems="10"
         android:inputType="textEmailAddress"
         android:padding="10dp"
-        android:text="tes"
         android:textColor="#626262"
+        android:textCursorDrawable="@drawable/cursorcolor"
         app:layout_constraintStart_toStartOf="@+id/textView3"
         app:layout_constraintTop_toBottomOf="@+id/textView3" />
 
-    <EditText
+    <com.google.android.material.textfield.TextInputLayout
         android:id="@+id/editTextTextPassword"
         android:layout_width="375dp"
         android:layout_height="48dp"
-        android:layout_marginTop="10dp"
         android:background="@drawable/border_inputtext"
-        android:drawableEnd="@drawable/baseline_remove_red_eye_24"
-        android:ems="10"
-        android:inputType="textPassword"
-        android:padding="10dp"
-        android:text="tes"
-        android:textColor="#000000"
+        app:boxStrokeWidth="0dp"
+        app:boxStrokeWidthFocused="0dp"
         app:layout_constraintStart_toStartOf="@+id/textView5"
-        app:layout_constraintTop_toBottomOf="@+id/textView5" />
+        app:layout_constraintTop_toBottomOf="@+id/textView5"
+        app:passwordToggleEnabled="true"
+        app:passwordToggleTint="#000000">
 
-    <EditText
+        <com.google.android.material.textfield.TextInputEditText
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            android:background="@drawable/bg_null"
+            android:ems="10"
+            android:gravity="center|start|left"
+            android:inputType="textPassword"
+            android:paddingTop="0dp"
+            android:paddingBottom="0dp"
+            android:textColor="#626262" />
+    </com.google.android.material.textfield.TextInputLayout>
+
+    <com.google.android.material.textfield.TextInputLayout
         android:id="@+id/editTextTextPassword2"
         android:layout_width="375dp"
         android:layout_height="48dp"
-        android:layout_marginTop="10dp"
         android:background="@drawable/border_inputtext"
-        android:drawableEnd="@drawable/baseline_remove_red_eye_24"
-        android:ems="10"
-        android:inputType="textPassword"
-        android:padding="10dp"
-        android:text="tes"
-        android:textColor="#000000"
+        app:boxStrokeWidth="0dp"
+        app:boxStrokeWidthFocused="0dp"
         app:layout_constraintStart_toStartOf="@+id/textView4"
-        app:layout_constraintTop_toBottomOf="@+id/textView4" />
+        app:layout_constraintTop_toBottomOf="@+id/textView4"
+        app:passwordToggleEnabled="true"
+        app:passwordToggleTint="#000000">
+
+        <com.google.android.material.textfield.TextInputEditText
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            android:background="@drawable/bg_null"
+            android:ems="10"
+            android:gravity="center|start|left"
+            android:inputType="textPassword"
+            android:paddingTop="0dp"
+            android:paddingBottom="0dp"
+            android:textColor="#626262" />
+    </com.google.android.material.textfield.TextInputLayout>
+
+
+
 
     <Button
         android:id="@+id/button"
@@ -147,7 +168,7 @@
         android:layout_marginTop="30dp"
         android:backgroundTint="#FF9F29"
         android:text="Button"
-        android:textColor="#222222"
+        android:textColor="#000000"
         android:textStyle="bold"
         app:cornerRadius="15dp"
         app:layout_constraintStart_toStartOf="@+id/textView6"
