+++ app/src/main/res/values/strings.xml
@@ -1,5 +1,7 @@
 <resources>
     <string name="app_name">Exless</string>
     <string name="ketentuan_layanan">Dengan membuat akun berarti kamu setuju dengan <b><u>Ketentuan Layanan Kami</u></b></string>
-    <string name="daftarlogin"><u>Udah daftar ? Login disini</u></string>
+    <string name="daftarlogin"><u>Sudah punya akun? Login disini</u></string>
+    <string name="akanregister"><u>Belum punya akun? Daftar</u></string>
+
 </resources>
